Working on commands

+++ commands/uim/commands/classes/commands/plugins/assets/symlink.d
@@ -22,7 +22,7 @@ class DPluginAssetsSymlinkCommand : DCommand {
         return "plugin-assets-symlink";
     }
 
-    override int execute(Json[string] arguments, IConsoleIo consoleIo) {
+    override ulong execute(Json[string] arguments, IConsoleIo consoleIo) {
         return super.execute(arguments, consoleIo);
     }
 
@@ -49,7 +49,7 @@ class DPluginAssetsSymlinkCommand : DCommand {
      * fallbacks to copying the assets. For vendor namespaced plugin, parent folder
      * for vendor name are created if required.
      */
-    override int execute(Json[string] arguments, IConsole aConsole = null) {
+    override ulong execute(Json[string] arguments, IConsole aConsole = null) {
         _io = aConsoleIo;
         _args = commandArguments;
